requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
uwsgi
psycopg2
pillow
JSON
scikit-image
tensorflow
keras
